add delwarn

+++ src/commands/removeWarn.ts
@@ -25,11 +25,10 @@ const command: commandHandler = {
                 guildId: interaction.guild!.id
             })
 
-            await Warning.findOne({
-                userId: member?.id,
-                guildId: interaction.guild!.id,
-                _id: getMemberWarns[0]._id
-            }).deleteOne();
+            if (getMemberWarns.length === 0) {
+                return void interaction.reply({ content: `${user.tag} has no warns.`, ephemeral: true });
+            }
+            await Warning.deleteOne({ userId: member?.id, guildId: interaction.guild!.id});
         } catch (err) {
             console.error(err);
             return void interaction.reply({ content: "❌ Failed to retrieve warns for the member.", ephemeral: true });
